Common | Domain 계층 UseCase 클래스들 네이밍 규칙 적용
