+++ app/src/main/java/com/applemango/runnerbe/domain/usecase/bookmark/GetBookmarkedPostsUseCase.kt
@@ -1,4 +1,4 @@
-package com.applemango.runnerbe.domain.usecase.post
+package com.applemango.runnerbe.domain.usecase.bookmark
 
 import com.applemango.runnerbe.domain.repository.PostRepository
 import com.applemango.runnerbe.presentation.state.CommonResponse
@@ -6,13 +6,14 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
 
-class PostReportUseCase @Inject constructor(
-    private val repo: PostRepository
-) {
-    operator fun invoke(postId: Int, userId : Int):Flow<CommonResponse> = flow {
+/**
+ * 북마크 게시글 목록 조회
+ */
+class GetBookmarkedPostsUseCase @Inject constructor(private val repo: PostRepository) {
+    operator fun invoke(userId: Int): Flow<CommonResponse> = flow {
         runCatching {
             emit(CommonResponse.Loading)
-            repo.reportPost(postId, userId)
+            repo.getBookmarkList(userId)
         }.onSuccess {
             emit(it)
         }.onFailure {
