+++ app/src/main/java/com/applemango/runnerbe/domain/usecase/post/SearchAddressByKeywordUseCase.kt
@@ -1,18 +1,17 @@
 package com.applemango.runnerbe.domain.usecase.post
 
-import android.util.Log
 import androidx.paging.Pager
 import androidx.paging.PagingConfig
-import com.applemango.runnerbe.data.network.response.BaseResponse
-import com.applemango.runnerbe.data.network.response.KakaoLocalResponse
 import com.applemango.runnerbe.data.paging.AddressSearchPagingSource
-import com.applemango.runnerbe.domain.repository.PostRepository
 import com.applemango.runnerbe.presentation.state.CommonResponse
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
 
-class GetAddressListUseCase @Inject constructor(
+/**
+ * 주소 검색 (카카오 API)
+ */
+class SearchAddressByKeywordUseCase @Inject constructor(
     private val addressSearchPagingSource: AddressSearchPagingSource
 ) {
     operator fun invoke(keyword: String): Flow<CommonResponse> = flow {
