+++ app/src/main/java/com/applemango/runnerbe/domain/usecase/runninglog/GetJoinedRunnersUseCase.kt
@@ -1,18 +1,17 @@
 package com.applemango.runnerbe.domain.usecase.runninglog
 
-import android.util.Log
 import com.applemango.runnerbe.domain.repository.RunningLogRepository
 import com.applemango.runnerbe.presentation.state.CommonResponse
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
 
-class GetMonthlyRunningLogListUseCase @Inject constructor(
+class GetJoinedRunnersUseCase @Inject constructor(
     private val runningLogRepository: RunningLogRepository
 ) {
-    suspend operator fun invoke(userId: Int, year: Int, month: Int): Flow<CommonResponse> = flow {
+    suspend operator fun invoke(userId: Int, gatheringId: Int): Flow<CommonResponse> = flow {
         kotlin.runCatching {
-            runningLogRepository.getMonthlyRunningLogList(userId, year, month)
+            runningLogRepository.getJoinedRunnerList(userId, gatheringId)
         }.onSuccess {
             emit(it)
         }.onFailure {
