+++ app/src/main/java/com/applemango/runnerbe/domain/usecase/runninglog/GetMonthlyRunningLogsUseCase.kt
@@ -1,23 +1,22 @@
 package com.applemango.runnerbe.domain.usecase.runninglog
 
-import com.applemango.runnerbe.data.network.request.RunningLogRequest
 import com.applemango.runnerbe.domain.repository.RunningLogRepository
 import com.applemango.runnerbe.presentation.state.CommonResponse
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
 
-class PatchRunningLogUseCase @Inject constructor(
+class GetMonthlyRunningLogsUseCase @Inject constructor(
     private val runningLogRepository: RunningLogRepository
 ) {
-    suspend operator fun invoke(userId: Int, logId: Int, runningLog: RunningLogRequest): Flow<CommonResponse> = flow {
+    suspend operator fun invoke(userId: Int, year: Int, month: Int): Flow<CommonResponse> = flow {
         kotlin.runCatching {
-            runningLogRepository.patchRunningLog(userId, logId, runningLog)
+            runningLogRepository.getMonthlyRunningLogList(userId, year, month)
         }.onSuccess {
             emit(it)
         }.onFailure {
             it.printStackTrace()
-            emit(CommonResponse.Failed(999, it.message?:"error"))
+            emit(CommonResponse.Failed.getDefaultFailed(it.message))
         }
     }
 }
