+++ app/src/main/java/com/applemango/runnerbe/domain/usecase/runninglog/UpdateRunningLogUseCase.kt
@@ -1,22 +1,23 @@
 package com.applemango.runnerbe.domain.usecase.runninglog
 
+import com.applemango.runnerbe.data.network.request.RunningLogRequest
 import com.applemango.runnerbe.domain.repository.RunningLogRepository
 import com.applemango.runnerbe.presentation.state.CommonResponse
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
 
-class GetJoinedRunnerListUseCase @Inject constructor(
+class UpdateRunningLogUseCase @Inject constructor(
     private val runningLogRepository: RunningLogRepository
 ) {
-    suspend operator fun invoke(userId: Int, gatheringId: Int): Flow<CommonResponse> = flow {
+    suspend operator fun invoke(userId: Int, logId: Int, runningLog: RunningLogRequest): Flow<CommonResponse> = flow {
         kotlin.runCatching {
-            runningLogRepository.getJoinedRunnerList(userId, gatheringId)
+            runningLogRepository.patchRunningLog(userId, logId, runningLog)
         }.onSuccess {
             emit(it)
         }.onFailure {
             it.printStackTrace()
-            emit(CommonResponse.Failed.getDefaultFailed(it.message))
+            emit(CommonResponse.Failed(999, it.message?:"error"))
         }
     }
 }
