+++ app/src/main/java/com/applemango/runnerbe/domain/usecase/runningtalk/ReportMessageUseCase.kt
@@ -2,20 +2,22 @@ package com.applemango.runnerbe.domain.usecase.runningtalk
 
 import com.applemango.runnerbe.domain.repository.RunningTalkRepository
 import com.applemango.runnerbe.presentation.state.CommonResponse
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
 
-class MessageSendUseCase @Inject constructor(
+class ReportMessageUseCase @Inject constructor(
     private val repo: RunningTalkRepository
 ) {
-    suspend operator fun invoke(roomId: Int, content: String?, url: String?): CommonResponse {
+    operator fun invoke(messageIdList : List<Int>) : Flow<CommonResponse> = flow {
         runCatching {
-            repo.sendMessage(roomId, content, url)
+            emit(CommonResponse.Loading)
+            repo.reportMessage(messageIdList)
         }.onSuccess {
-            return it
+            emit(it)
         }.onFailure {
             it.printStackTrace()
-            return CommonResponse.Failed(999, it.message?:"error")
+            emit(CommonResponse.Failed(999, it.message?:"error"))
         }
-        return CommonResponse.Failed(999, "error")
     }
 }
