+++ app/src/main/java/com/applemango/runnerbe/domain/usecase/user/GetOtherUserDataUseCase.kt
@@ -1,4 +1,4 @@
-package com.applemango.runnerbe.domain.usecase.runninglog
+package com.applemango.runnerbe.domain.usecase.user
 
 import com.applemango.runnerbe.domain.repository.RunningLogRepository
 import com.applemango.runnerbe.presentation.state.CommonResponse
@@ -6,7 +6,7 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
 
-class GetOtherUserProfileUseCase @Inject constructor(
+class GetOtherUserDataUseCase @Inject constructor(
     private val runningLogRepository: RunningLogRepository
 ) {
     suspend operator fun invoke(targetUserId: Int): Flow<CommonResponse> = flow {
