+++ app/src/main/java/com/applemango/runnerbe/domain/usecase/user/RegisterUserUseCase.kt
@@ -1,19 +1,20 @@
-package com.applemango.runnerbe.domain.usecase
+package com.applemango.runnerbe.domain.usecase.user
 
+import com.applemango.runnerbe.data.network.request.JoinUserRequest
 import com.applemango.runnerbe.domain.repository.UserRepository
 import com.applemango.runnerbe.presentation.state.CommonResponse
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
 
-class JobChangeUseCase @Inject constructor(
-    private val repo : UserRepository
+class RegisterUserUseCase @Inject constructor(
+    private val repo: UserRepository
 ) {
 
-    operator fun invoke(userId: Int, job: String) : Flow<CommonResponse> = flow  {
+    operator fun invoke(request : JoinUserRequest) : Flow<CommonResponse> = flow {
         runCatching {
             emit(CommonResponse.Loading)
-            repo.jobChange(userId, job)
+            repo.joinUser(request)
         }.onSuccess {
             emit(it)
         }.onFailure { e ->
