+++ app/src/main/java/com/applemango/runnerbe/domain/usecase/user/UpdateJobUseCase.kt
@@ -1,4 +1,4 @@
-package com.applemango.runnerbe.domain.usecase.myinfo
+package com.applemango.runnerbe.domain.usecase.user
 
 import com.applemango.runnerbe.domain.repository.UserRepository
 import com.applemango.runnerbe.presentation.state.CommonResponse
@@ -6,15 +6,17 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
 
-class NicknameChangeUseCase @Inject constructor(private val repo : UserRepository) {
+class UpdateJobUseCase @Inject constructor(
+    private val repo : UserRepository
+) {
 
-    operator fun invoke(userId: Int, nickname : String) : Flow<CommonResponse> = flow {
+    operator fun invoke(userId: Int, job: String) : Flow<CommonResponse> = flow  {
         runCatching {
             emit(CommonResponse.Loading)
-            repo.nicknameChange(userId, nickname)
+            repo.jobChange(userId, job)
         }.onSuccess {
             emit(it)
-        }.onFailure { e->
+        }.onFailure { e ->
             e.printStackTrace()
             emit(CommonResponse.Failed.getDefaultFailed(e.message))
         }
