+++ app/src/main/java/com/applemango/runnerbe/domain/usecase/user/UpdateUserPaceUseCase.kt
@@ -1,21 +1,21 @@
-package com.applemango.runnerbe.domain.usecase
+package com.applemango.runnerbe.domain.usecase.user
 
+import com.applemango.runnerbe.domain.entity.Pace
 import com.applemango.runnerbe.domain.repository.UserRepository
 import com.applemango.runnerbe.presentation.state.CommonResponse
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
 
-class PatchUserImageUseCase @Inject constructor(
-    private val repo: UserRepository
-) {
-    operator fun invoke(imageUrl : String?) : Flow<CommonResponse> = flow {
+class UpdateUserPaceUseCase @Inject constructor(private val repo : UserRepository)  {
+
+    operator fun invoke(userId: Int, pace: Pace)  : Flow<CommonResponse> = flow {
         runCatching {
             emit(CommonResponse.Loading)
-            repo.patchUserImage(imageUrl)
+            repo.patchUserPaceRegist(userId, pace.key)
         }.onSuccess {
             emit(it)
-        }.onFailure { e ->
+        }.onFailure { e->
             e.printStackTrace()
             emit(CommonResponse.Failed.getDefaultFailed(e.message))
         }
